Add test_risk_score_multiple_contact
Only add a single trace token if initial risk score is < 1 from an index

Remove trace_tokens at the end of tracing from individuals whose risk
score < 1

Consolidate removing of trace_tokens from an individual into a single
function

traced_on_this_trace is now a double since it is used as the cumulative
risk score per trace

+++ src/individual.h
@@ -52,7 +52,7 @@ struct individual{
 	
 	trace_token *trace_tokens;
 	trace_token *index_trace_token;
-	int traced_on_this_trace;
+	double traced_on_this_trace;
 
 	int app_user;
 };
